Revert "Merge remote-tracking branch 'origin/testbranch' into testbranch"
This reverts commit dadf64cb54b1fd31a2ac2cfe6ab110f4dbf5d798, reversing
changes made to 3caec7145a7965bab15b675e6e3969b5f98eba71.
