--- DogeClicker/app/src/main/java/com/example/dogeclicker/models/PC.java
@@ -1,4 +1,0 @@
-package com.example.dogeclicker.models;
-
-public class PC {
-}
