+++ DogeClicker/app/src/main/java/com/example/dogeclicker/views/MainActivity.java
@@ -1,11 +1,9 @@
-package com.example.dogeclicker.view;
+package com.example.dogeclicker.views;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
 
 import com.example.dogeclicker.R;
-
 
 public class MainActivity extends AppCompatActivity {
 
@@ -13,29 +11,5 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-    }
-
-
-
-    public void onUpgradeCilck(View v){
-        setContentView(R.layout.perm_upgrade_page);
-
-    }
-
-    public void onDogeCoinClick(View v){
-
-    }
-
-    public void onCursorClick(View v){
-
-
-    }
-
-    public void onCPUClick(View v){
-
-    }
-    public void onRAMClick(View v){
-
     }
 }
