--- DogeClicker/app/src/main/java/controller/ClickManager.java
@@ -1,5 +1,0 @@
-package controller;
-
-public class ClickManager {
-
-}
