+++ DogeClicker/app/src/main/java/models/Upgrade.java
@@ -1,38 +1,4 @@
 package models;
 
-import com.example.dogeclicker.models.UpgradeType;
-
 public class Upgrade {
-    private String name;
-    private UpgradeType type;
-    private double multiplier;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public UpgradeType getType() {
-        return type;
-    }
-
-    public void setType(UpgradeType type) {
-        this.type = type;
-    }
-
-    public double getMultiplier() {
-        return multiplier;
-    }
-
-    public void setMultiplier(double multiplier) {
-        this.multiplier = multiplier;
-    }
-
-    @Override
-    public String toString(){
-     return "uwu";
-    }
 }
