+++ DogeClicker/app/src/main/java/models/UpgradeType.java
@@ -1,4 +1,4 @@
-package com.example.dogeclicker.models;
+package models;
 
 public enum UpgradeType {
     BASIC,
