+++ DogeClicker/app/src/main/res/layout/activity_main.xml
@@ -7,71 +7,13 @@
     android:layout_height="match_parent"
     tools:context=".views.MainActivity">
 
-    <Button
-        android:id="@+id/ramBtn"
-        android:layout_width="133dp"
-        android:layout_height="158dp"
-        android:layout_marginStart="3dp"
-        android:layout_marginTop="578dp"
-        android:layout_marginBottom="1dp"
-        android:onClick="onRAMClick"
-        android:text="RAM"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/cpuBtn"
-        android:layout_width="133dp"
-        android:layout_height="158dp"
-        android:layout_marginStart="4dp"
-        android:layout_marginTop="573dp"
-        android:onClick="onCPUClick"
-        android:text="CPU"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/ramBtn"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0" />
-
-    <Button
-        android:id="@+id/cursorBtn"
-        android:layout_width="133dp"
-        android:layout_height="158dp"
-        android:layout_marginStart="5dp"
-        android:layout_marginTop="576dp"
-        android:layout_marginEnd="1dp"
-        android:onClick="onCursorClick"
-        android:text="Cursor"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/cpuBtn"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/dogeBtn"
-        android:layout_width="133dp"
-        android:layout_height="114dp"
-        android:layout_marginStart="140dp"
-        android:layout_marginEnd="138dp"
-        android:layout_marginBottom="191dp"
-        android:onClick="onDogeCoinClick"
-        android:text="Doge"
-        app:layout_constraintBottom_toTopOf="@+id/cpuBtn"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/permUpgradeBtn"
+    <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="311dp"
-        android:layout_marginTop="68dp"
-        android:layout_marginEnd="12dp"
-        android:layout_marginBottom="459dp"
-        android:onClick="onUpgradeCilck"
-        android:text="Upgrades"
-        app:layout_constraintBottom_toTopOf="@+id/cursorBtn"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
+        android:text="Hello World!"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
